Initial commit on inference
I have created an inference module, to help me predict data (and also test it) using a pre-trained model.

+++ src/utils/nifti.py
@@ -5,7 +5,7 @@ import os
 
 def save_nifti(segmentation_volume, output_file):
     # Create a NIfTI image object
-    nifti_img = nib.Nifti1Image(segmentation_volume, affine=np.eye(4))
+    nifti_img = nib.Nifti1Image(segmentation_volume.astype(np.int16), affine=np.eye(4))
 
     # Create the directory if it doesn't exist
     output_directory = os.path.dirname(output_file)
